trying loop transition

+++ src/main/java/com/example/ap_project_snakesandladders/Token.java
@@ -17,8 +17,8 @@ public class Token {
         this.status = status;
     }
 
-    private double x_coordinate;
-    private double y_coordinate;
+    private int x_coordinate;
+    private int y_coordinate;
 
     public ImageView getToken() {
         return this.token;
@@ -28,7 +28,7 @@ public class Token {
         this.token = token;
     }
 
-    public double getX_coordinate() {
+    public int getX_coordinate() {
         return this.x_coordinate;
     }
 
@@ -36,7 +36,7 @@ public class Token {
         this.x_coordinate = x_coordinate;
     }
 
-    public double getY_coordinate() {
+    public int getY_coordinate() {
         return this.y_coordinate;
     }
 
@@ -46,7 +46,7 @@ public class Token {
 
 
 
-    public Token(ImageView token, double x_coordinate, double y_coordinate) {
+    public Token(ImageView token, int x_coordinate, int y_coordinate) {
         this.token = token;
         this.x_coordinate = x_coordinate;
         this.y_coordinate = y_coordinate;
